+++ examples/ex-grid_editor_date.R
@@ -2,6 +2,7 @@ library(toastui)
 
 dat <- data.frame(
   date = Sys.Date() + 1:10,
+  date_locale = format(Sys.Date() + 1:10, format = "%d/%m/%Y"),
   month = format(Sys.Date() + 1:10, format = "%Y-%m"),
   year = format(Sys.Date() + 1:10, format = "%Y"),
   time1 = Sys.time() + 1:10,
@@ -9,27 +10,48 @@ dat <- data.frame(
 )
 
 
-datagrid(dat) %>% 
+datagrid(
+  data = dat,
+  datepicker_locale = list(
+    titles = list(
+      DD = c("Dimanche", "Lundi", "Mardi", "Mercredi", "Jeudi", "Vendredi", "Samedi"),
+      D = c("Dim", "Lun", "Mar", "Mer", "Jeu", "Ven", "Sam"),
+      MMMM = c("Janvier", "Février", "Mars", "Avril", "Mai", "Juin", "Juillet",
+               "Août", "Septembre", "Octobre", "Novembre", "Décembre"),
+      MMM = c("Jan", "Fév", "Mar", "Avr", "Mai", "Juin", "Juil", "Aou", "Sept", "Oct", "Nov", "Déc")
+    ),
+    titleFormat = "MMM yyyy",
+    todayFormat = "DD dd MMMM yyyy",
+    date = "Date",
+    time = "Heure"
+  )
+) %>%
   grid_editor_date(
     column = "date"
-  ) %>% 
+  )%>%
+  grid_editor_date(
+    column = "date_locale",
+    format = "dd/MM/yyyy",
+    language = "custom",
+    weekStartDay = "Mon"
+  ) %>%
   grid_editor_date(
     column = "month",
     type = "month",
     format = "yyyy-MM"
-  ) %>% 
+  ) %>%
   grid_editor_date(
     column = "year",
     type = "year",
     format = "yyyy"
-  ) %>% 
+  ) %>%
   grid_editor_date(
-    column = "time1", 
+    column = "time1",
     timepicker = "tab",
     format = "yyyy-MM-dd HH:mm"
-  ) %>% 
+  ) %>%
   grid_editor_date(
-    column = "time2", 
+    column = "time2",
     timepicker = "normal",
     format = "yyyy-MM-dd HH:mm"
   )